Add support for outdated/deprecated SSIDs
The app is now capable to show the user if a SSID is outdated/deprecated.
The app checks the ssids.json for a array nemed "deprecated", that holds
all freifunk ssids that are no longer official in use. A new option was
added so that the user can decide if outdated ssids should always be
shown. The default setting is to not always show outdated ssids. If the
app finds an outdated freifunk SSID in the network configuration, it
always shows it and sets its background color to orange. If the user
opted to always show outdated SSIDs, the background color of outdated
SSIDs will be set to yellow. The info activity was updated to inform about
the meaning of the different colors.